+++ packages/caelus/src/index.ts
@@ -23,3 +23,4 @@ export * from "./vedic.js";
 export * from "./directions.js";
 export * from "./vargas.js";
 export * from "./yogini.js";
+export * from "./yogas.js";